visitor: move `Delete unused variables` to a single file
